Move out vg concat
